+++ src/main/java/seedu/DueQuestException.java
@@ -1,0 +1,16 @@
+package seedu;
+/**
+ * DukeException is used to represent duke-specific exceptions that are likely
+ * to occur
+ */
+public class DueQuestException extends Exception {
+    DueQuestExceptionType exceptionType;
+
+    public DueQuestExceptionType getExceptionType() {
+        return exceptionType;
+    }
+
+    public DueQuestException(DueQuestExceptionType exceptionType){
+        this.exceptionType=exceptionType;
+    }
+}
