+++ src/main/java/seedu/DueQuestExceptionType.java
@@ -2,6 +2,6 @@ package seedu;
 /**
  * DukeExceptionType represents the different kind of Duke-specific Exceptions
  */
-public enum DukeExceptionType {
+public enum DueQuestExceptionType {
     MISSING_DESCRIPTION,WRONG_INPUT_FORMAT,MISSING_DEADLINE, MISSING_EVENT_INFO, INVALID_COMMAND
 }
