+++ src/test/java/seedu/duequest/DueQuestTest.java
@@ -1,10 +1,10 @@
-package seedu.duke;
+package seedu.duequest;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-class DukeTest {
+class DueQuestTest {
     @Test
     public void sampleTest() {
         assertTrue(true);
